Fix to always show separation between hp and shield
Fix to always show separation between hp and shield (even when player
doesn't have a shield). This might need to be changed later so it
automatically scales the HP if the player has no shield. However this is
only in the very early stages of the game. The player receives a shield
after the first few quests, so low priority and perhaps should not be
done at all.