Workaround for self.eChatChannel is nil lua error
In some cases - somehow - the self.eChatChannel (which comes from
ChatSystemLib) receives a nil value. Perhaps caused by a different
addon? Either way, this is a workaround for the lua error that might
occur when the addon tries to send something to the chat when
self.eChatChannel has a nil value.